Program updated

+++ almost_equal_sum.cpp
@@ -6,15 +6,15 @@ string output_file = "./out.txt";
 
 struct SetInfo {
     vector<string> elements;
-    vector<int> values;
-    int sum;
+    vector<float> values;
+    float sum;
 };
 
-vector<SetInfo> divide_almost_equally(vector<pair<string, int>>& inputs, int num_chunks) {
-    sort(inputs.begin(), inputs.end(), [](const pair<string, int>& a, const pair<string, int>& b) {
+vector<SetInfo> divide_almost_equally(vector<pair<string, float>>& inputs, int num_chunks) {
+    sort(inputs.begin(), inputs.end(), [](const pair<string, float>& a, const pair<string, float>& b) {
         return a.second > b.second;
     });
-    priority_queue<pair<int, int>, vector<pair<int, int>>, greater<pair<int, int>>> heap;
+    priority_queue<pair<float, int>, vector<pair<float, int>>, greater<pair<float, int>>> heap;
     for (int idx = 0; idx < num_chunks; ++idx) {
         heap.push(make_pair(0, idx));
     }
@@ -22,7 +22,7 @@ vector<SetInfo> divide_almost_equally(vector<pair<string, int>>& inputs, int num
     int arr_idx = 0;
     while (arr_idx < inputs.size()) {
         auto top = heap.top();
-        int set_sum = top.first;
+        float set_sum = top.first;
         int set_idx = top.second;
         heap.pop();
         sets[set_idx].elements.push_back(inputs[arr_idx].first);
@@ -40,11 +40,11 @@ vector<SetInfo> divide_almost_equally(vector<pair<string, int>>& inputs, int num
 //     cout << endl;
 // }
 
-void printBefore(vector<pair<string,int>> inputs){
+void printBefore(vector<pair<string,float>> inputs){
 
     ofstream outputFile(output_file);
 
-    int total_sum = accumulate(inputs.begin(), inputs.end(), 0, [](int sum, const pair<string, int>& p) {
+    float total_sum = accumulate(inputs.begin(), inputs.end(), 0, [](float sum, const pair<string, float>& p) {
         return sum + p.second;
     });    
     
@@ -52,7 +52,7 @@ void printBefore(vector<pair<string,int>> inputs){
     outputFile << endl;
 
     outputFile  << "Sorted Array: ";
-    for (pair<string,int> val : inputs) {
+    for (pair<string,float> val : inputs) {
         outputFile  << val.second << " ";
     }
     outputFile  << endl << "Total sum: " << total_sum <<  endl;
@@ -78,7 +78,7 @@ void printAfter(vector<SetInfo> sets){
         outputFile << endl;
 
         outputFile << "Values ";
-        for (int num : sets[i].values) {
+        for (float num : sets[i].values) {
             outputFile << num << " ";
         }
         outputFile << endl;
@@ -103,9 +103,9 @@ int main(int argc, char* argv[]) {
     }
 
     // Reading inputs from input_file
-    vector<pair<string,int>> inputs;
+    vector<pair<string,float>> inputs;
     string name;
-    int value;
+    float value;
     inputFile >> num_chunks;
     while (inputFile >> name >> value) { 
         inputs.push_back(make_pair(name, value));
@@ -113,7 +113,7 @@ int main(int argc, char* argv[]) {
     inputFile.close(); 
 
     // Counting total sum before process
-    sort(inputs.begin(), inputs.end(), [](const pair<string, int>& a, const pair<string, int>& b) {
+    sort(inputs.begin(), inputs.end(), [](const pair<string, float>& a, const pair<string, float>& b) {
         return a.second > b.second;
     });    
     // printdummy();
